+++ RubiChess/board.cpp
@@ -438,7 +438,7 @@ bool chessposition::applyMove(string s)
 
 
 template <MoveType Mt>
-void evaluateMoves(chessmovelist *ml, chessposition *pos, int32_t **cmptr)
+void evaluateMoves(chessmovelist *ml, chessposition *pos, int16_t **cmptr)
 {
     for (int i = 0; i < ml->length; i++)
     {
